+++ dsh_sdk/src/schema_store/client.rs
@@ -1,7 +1,7 @@
 use super::api::SchemaStoreApi;
 use super::request::Request;
 use super::types::*;
-use super::{Result, SchemaStoreError};
+use super::Result;
 use crate::Dsh;
 
 /// High level Schema Store Client
@@ -35,21 +35,24 @@ where
     /// ## Returns
     /// Returns a Result of the compatibility level of given subject
     ///
+    /// ## Arguments
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
+    ///
     /// ## Example
     /// ```no_run
     /// use dsh_sdk::schema_store::SchemaStoreClient;
+    /// use dsh_sdk::schema_store::types::SubjectName;
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
-    /// println!("Config: {:?}", client.subject_compatibility("scratch.example-topic.tenant-value").await);
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// println!("Config: {:?}", client.subject_compatibility(&subject_name).await);
+    /// # Ok(())
     /// # }
     ///
-    pub async fn subject_compatibility<Sn>(&self, subject: Sn) -> Result<Compatibility>
-    where
-        Sn: Into<SubjectName>,
-    {
-        Ok(self.get_config_subject(subject.into().name()).await?.into())
+    pub async fn subject_compatibility(&self, subject: &SubjectName) -> Result<Compatibility> {
+        Ok(self.get_config_subject(subject.name()).await?.into())
     }
 
     /// Set the compatibility level for a subject
@@ -57,32 +60,32 @@ where
     /// Set compatibility on subject level. With 1 schema stored in the subject, you can change it to any compatibility level.
     /// Else, you can only change into a less restrictive level.
     ///
+    /// ## Arguments
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
+    ///
     /// ## Returns
     /// Returns a Result of the new compatibility level
     ///
     /// ## Example
     /// ```no_run
     /// use dsh_sdk::schema_store::SchemaStoreClient;
-    /// use dsh_sdk::schema_store::types::Compatibility;
+    /// use dsh_sdk::schema_store::types::{Compatibility, SubjectName};
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
-    /// client.subject_compatibility_update("scratch.example-topic.tenant-value", Compatibility::FULL).await.unwrap();
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// client.subject_compatibility_update(&subject_name, Compatibility::FULL).await?;
+    /// # Ok(())
     /// # }
     /// ```
-    ///
-    /// TODO: untested as this API method does not seem to work at all on DSH
-    pub async fn subject_compatibility_update<Sn>(
+    pub async fn subject_compatibility_update(
         &self,
-        subject: Sn,
+        subject: &SubjectName,
         compatibility: Compatibility,
-    ) -> Result<Compatibility>
-    where
-        Sn: Into<SubjectName>,
-    {
+    ) -> Result<Compatibility> {
         Ok(self
-            .put_config_subject(subject.into().name(), compatibility)
+            .put_config_subject(subject.name(), compatibility)
             .await?
             .into())
     }
@@ -114,19 +117,18 @@ where
     /// ## Example
     /// ```no_run
     /// use dsh_sdk::schema_store::SchemaStoreClient;
+    /// use dsh_sdk::schema_store::types::SubjectName;
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
-    /// println!("Available versions: {:?}", client.subject_versions("scratch.example-topic.tenant-value").await);
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// println!("Available versions: {:?}", client.subject_versions(&subject_name).await);
+    /// # Ok(())
     /// # }
     /// ```
-    pub async fn subject_versions<Sn>(&self, subject: Sn) -> Result<Vec<i32>>
-    where
-        Sn: Into<SubjectName>,
-    {
-        self.get_subjects_subject_versions(subject.into().name())
-            .await
+    pub async fn subject_versions(&self, subject: &SubjectName) -> Result<Vec<i32>> {
+        self.get_subjects_subject_versions(subject.name()).await
     }
 
     /// Get subject for specific version
@@ -140,25 +142,25 @@ where
     /// use dsh_sdk::schema_store::types::{SubjectName, SubjectVersion};
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
     ///
     /// // Get the latest version of the schema
-    /// let subject = client.subject(SubjectName::TopicNameStrategy{topic: "scratch.example-topic.tenant".to_string(), key: false}, SubjectVersion::Latest).await.unwrap();
+    /// let subject = client.subject(&subject_name, SubjectVersion::Latest).await?;
     /// let raw_schema = subject.schema;
     ///
     /// // Get a specific version of the schema
-    /// let subject = client.subject("scratch.example-topic.tenant-value", SubjectVersion::Version(1)).await.unwrap();
+    /// let subject = client.subject(&subject_name, SubjectVersion::Version(1)).await?;
     /// let raw_schema = subject.schema;
+    /// # Ok(())
     /// # }
     /// ```
-    pub async fn subject<Sn, V>(&self, subject: Sn, version: V) -> Result<Subject>
+    pub async fn subject<V>(&self, subject: &SubjectName, version: V) -> Result<Subject>
     where
-        //Sn: TryInto<SubjectName, Error = SchemaStoreError>,
-        Sn: Into<SubjectName>,
         V: Into<SubjectVersion>,
     {
-        let subject = subject.into().name();
+        let subject = subject.name();
         let version = version.into();
         self.get_subjects_subject_versions_id(subject, version.to_string())
             .await
@@ -167,8 +169,8 @@ where
     /// Get the raw schema string for the specified version of subject.
     ///
     /// ## Arguments
-    /// - `subject`: Anything that can be converted into a [SubjectName] (Returns error if invalid SubjectStrategy)
-    /// - `schema`: Anything that can be converted into a [RawSchemaWithType]
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
+    /// - `schema`: [RawSchemaWithType], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SchemaType)
     ///
     /// ## Returns
     /// Returns a Result of the raw schema string for the given subject and version
@@ -176,29 +178,28 @@ where
     /// ## Example
     /// ```no_run
     /// use dsh_sdk::schema_store::SchemaStoreClient;
+    /// use dsh_sdk::schema_store::types::SubjectName;
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
-    /// let raw_schema = client.subject_raw_schema("scratch.example-topic.tenant-value", 1).await.unwrap();
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// let raw_schema = client.subject_raw_schema(&subject_name, 1).await.unwrap();
+    /// # Ok(())
     /// # }
     /// ```
-    pub async fn subject_raw_schema<Sn, V>(&self, subject: Sn, version: V) -> Result<String>
+    pub async fn subject_raw_schema<V>(&self, subject: &SubjectName, version: V) -> Result<String>
     where
-        Sn: Into<SubjectName>,
         V: Into<SubjectVersion>,
     {
-        self.get_subjects_subject_versions_id_schema(
-            subject.into().name(),
-            version.into().to_string(),
-        )
-        .await
+        self.get_subjects_subject_versions_id_schema(subject.name(), version.into().to_string())
+            .await
     }
 
     /// Get all schemas for a subject
     ///    
     /// ## Arguments
-    /// - `subject`: Anything that can be converted into a [SubjectName] (Returns error if invalid SubjectStrategy)
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
     ///
     /// ## Returns
     /// Returns a Result of all schemas for the given subject
@@ -209,22 +210,37 @@ where
     /// use dsh_sdk::schema_store::types::SubjectName;
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
-    /// let subjects = client.subject_all_schemas(SubjectName::TopicNameStrategy{topic: "scratch.example-topic.tenant".to_string(), key: false}).await.unwrap();
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// let subjects = client.subject_all_schemas(&subject_name).await?;
+    /// # Ok(())
     /// # }
-    pub async fn subject_all_schemas<Sn>(&self, subject: Sn) -> Result<Vec<Subject>>
-    where
-        Sn: Into<SubjectName> + Clone,
-    {
-        let versions = self.subject_versions(subject.clone()).await?;
+    pub async fn subject_all_schemas(&self, subject: &SubjectName) -> Result<Vec<Subject>> {
+        let versions = self.subject_versions(&subject).await?;
         let mut subjects = Vec::new();
         for version in versions {
-            let subject = self.subject(subject.clone(), version).await?;
+            let subject = self.subject(&subject, version).await?;
             subjects.push(subject);
         }
         Ok(subjects)
     }
+
+    /// Get all schemas for a topic
+    ///
+    /// ## Arguments
+    /// - `topic`: &str/String of the topic name
+    ///
+    /// ## Returns
+    ///
+    // pub async fn topic_all_schemas<S>(&self, topic: S) -> Result<(Vec<Subject>,Vec<Subject>)>
+    // where
+    //     S: AsRef<str>,
+    // {
+    //     let key_schemas = self.subject_all_schemas((topic.as_ref(), true)).await?;
+    //     let value_schemas = self.subject_all_schemas((topic.as_ref(), false)).await?;
+    //     Ok(subjects)
+    // }
 
     /// Post a new schema for a (new) subject
     ///
@@ -236,8 +252,8 @@ where
     ///     - schema is invalid
     ///
     /// ## Arguments
-    /// - `subject`: Anything that can be converted into a [SubjectName] (Returns error if invalid SubjectStrategy)
-    /// - `schema`: Anything that can be converted into a [RawSchemaWithType]
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
+    /// - `schema`: [RawSchemaWithType], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SchemaType)
     ///
     /// ## Returns
     /// Returns a Result of the new schema ID.
@@ -245,29 +261,34 @@ where
     ///
     /// ## Example
     /// ```no_run
-    /// use dsh_sdk::schema_store::{SchemaStoreClient, types::SchemaType};
+    /// use dsh_sdk::schema_store::SchemaStoreClient;
+    /// use dsh_sdk::schema_store::types::{RawSchemaWithType, SubjectName};
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
+    ///
+    /// // Get subjectname (note it ends on "-value")
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
     ///
     /// // You can provide the schema as a raw string (Schema type is optional, it will be detected automatically)
     /// let raw_schema = r#"{ "type": "record", "name": "User", "fields": [ { "name": "name", "type": "string" } ] }"#;
-    /// let schema_version = client.subject_add_schema("scratch.example-topic.tenant-value", (raw_schema, SchemaType::AVRO)).await.unwrap();
+    /// let schema_with_type:RawSchemaWithType = raw_schema.try_into()?;
+    /// let schema_version = client.subject_add_schema(&subject_name, schema_with_type).await?;
     ///
     /// // Or if you have a schema object
-    /// let avro_schema = apache_avro::Schema::parse_str(raw_schema).unwrap(); // or ProtoBuf or JSON schema
-    /// let schema_version = client.subject_add_schema("scratch.example-topic.tenant-value", avro_schema).await.unwrap();
+    /// let avro_schema:RawSchemaWithType  = apache_avro::Schema::parse_str(raw_schema)?.try_into()?; // or ProtoBuf or JSON schema
+    /// let schema_version = client.subject_add_schema(&subject_name, avro_schema).await?;
+    /// # Ok(())
     /// # }
     /// ```
-    pub async fn subject_add_schema<Sn, Sc>(&self, subject: Sn, schema: Sc) -> Result<i32>
-    where
-        Sn: Into<SubjectName>,
-        Sc: TryInto<RawSchemaWithType, Error = SchemaStoreError>,
-    {
-        let schema = schema.try_into()?;
+    pub async fn subject_add_schema(
+        &self,
+        subject: &SubjectName,
+        schema: RawSchemaWithType,
+    ) -> Result<i32> {
         Ok(self
-            .post_subjects_subject_versions(subject.into().name(), schema)
+            .post_subjects_subject_versions(subject.name(), schema)
             .await?
             .id())
     }
@@ -283,8 +304,8 @@ where
     ///     - schema is invalid
     ///
     /// ## Arguments
-    /// - `subject`: Anything that can be converted into a [SubjectName] (Returns error if invalid SubjectStrategy)
-    /// - `schema`: Anything that can be converted into a [RawSchemaWithType]
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
+    /// - `schema`: [RawSchemaWithType], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SchemaType)
     ///
     /// ## Returns
     /// If schema exists, it will return with the existing version and schema ID.
@@ -292,25 +313,25 @@ where
     /// ## Example
     /// ```no_run
     /// use dsh_sdk::schema_store::SchemaStoreClient;
-    /// use dsh_sdk::schema_store::types::{SubjectName, SchemaType};
+    /// use dsh_sdk::schema_store::types::{SubjectName, SchemaType, RawSchemaWithType};
     ///
     /// # #[tokio::main]
-    /// # async fn main() {
+    /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// let client = SchemaStoreClient::new();
     ///
     /// // You can provide the schema as a raw string (Schema type is optional, it will be detected automatically)
-    /// let raw_schema = r#"{ "type": "record", "name": "User", "fields": [ { "name": "name", "type": "string" } ] }"#;
-    /// let subject = client.subject_schema_exist("scratch.example-topic.tenant-value", (raw_schema, SchemaType::AVRO)).await.unwrap();
+    /// let raw_schema: RawSchemaWithType = r#"{ "type": "record", "name": "User", "fields": [ { "name": "name", "type": "string" } ] }"#.try_into()?;
+    /// let subject_name: SubjectName = "scratch.example-topic.tenant-value".try_into()?;
+    /// let subject = client.subject_schema_exist(&subject_name, raw_schema).await?;
+    /// # Ok(())
     /// # }
     /// ```
-    pub async fn subject_schema_exist<Sn, Sc>(&self, subject: Sn, schema: Sc) -> Result<Subject>
-    where
-        Sn: Into<SubjectName>,
-        Sc: TryInto<RawSchemaWithType, Error = SchemaStoreError>,
-    {
-        let schema = schema.try_into()?;
-        self.post_subjects_subject(subject.into().name(), schema)
-            .await
+    pub async fn subject_schema_exist(
+        &self,
+        subject: &SubjectName,
+        schema: RawSchemaWithType,
+    ) -> Result<Subject> {
+        self.post_subjects_subject(subject.name(), schema).await
     }
 
     /// Check if schema is compatible with a specific version of a subject based on the compatibility level
@@ -319,27 +340,24 @@ where
     /// If this subject’s compatibility level was never changed, then the global compatibility level applies.
     ///
     /// ## Arguments
-    /// - `subject`: Anything that can be converted into a [SubjectName] (Returns error if invalid SubjectStrategy)
+    /// - `subject`: [SubjectName], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SubjectStrategy)
     /// - `version`: Anything that can be converted into a [SubjectVersion]
-    /// - `schema`: Anything that can be converted into a [RawSchemaWithType]
+    /// - `schema`: [RawSchemaWithType], use [TryInto] to convert from &str/String (Returns [SchemaStoreError] error if invalid SchemaType)
     ///
     /// ## Returns
     /// Returns a Result of a boolean if the schema is compatible with the given version of the subject
-    pub async fn subject_new_schema_compatibility<Sn, Sv, Sc>(
+    pub async fn subject_new_schema_compatibility<Sv>(
         &self,
-        subject: Sn,
+        subject: &SubjectName,
         version: Sv,
-        schema: Sc,
+        schema: RawSchemaWithType,
     ) -> Result<bool>
     where
-        Sn: Into<SubjectName>,
         Sv: Into<SubjectVersion>,
-        Sc: TryInto<RawSchemaWithType, Error = SchemaStoreError>,
     {
-        let schema = schema.try_into()?;
         Ok(self
             .post_compatibility_subjects_subject_versions_id(
-                subject.into().name(),
+                subject.name(),
                 version.into().to_string(),
                 schema,
             )