+++ dsh_sdk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - unreleased
 ### Added
+- DSH Kafka Config trait to configure kafka client with RDKafka implementation
+- DSH Schema store API Client
 - New public functions `dsh_sdk::certificates::Cert`
   - Bootstrap to DSH
   - Read certificates from PKI_CONFIG_DIR
@@ -25,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `dsh_sdk::dsh::certificates` to `dsh_sdk::certificates`
   - Private module `dsh_sdk::dsh::bootstrap` and `dsh_sdk::dsh::pki_config_dir` are now part of `certificates` module
 - Moved `dsh_sdk::mqtt_token_fetcher` to `dsh_sdk::protocol_adapters::token_fetcher` where it is renamed to `ProtocolTokenFetcher`
-  - **NOTE** Cargo.toml feature flag falls now under `mqtt-protocol` (`mqtt_token_fetcher` will be removed in v0.6.0)
+  - **NOTE** Cargo.toml feature flag falls now under `protocol-token-fetcher` (`mqtt-token-fetcher` will be removed in v0.6.0)
 - Moved `dsh_sdk::dlq` to `dsh_sdk::utils::dlq` 
 - Moved `dsh_sdk::graceful_shutdown` to `dsh_sdk::utils::graceful_shutdown`
 - Moved `dsh_sdk::metrics` to `dsh_sdk::utils::metrics`